Make date normalization part of spreadsheet standardizer
## Description

Move this logic out of DataRefiner and into SpreadsheetStandardizer.
The dates from spreadsheets are normalized on the way into a
CollectedIncentive now, which will let us automatically clean up the
existing dates in sheets as part of JSON-back-to-sheet flows.

The `_raw` versions of the fields are gone, and the two date fields
are now pass-through. They're parsed and reformatted similarly to how
dollar-amount fields are. If they're unparseable, they're left as-is
so that ajv validation will fail, and whoever is doing the conversion
can go fix the spreadsheet manually.

## Test Plan

`yarn lint` and `yarn test`. `yarn check-spreadsheet-health` passes
except the NV sheet, but all the diffs there are part of the
currently-outstanding #358.
