+++ scripts/lib/spreadsheet-mappings.ts
@@ -20,8 +20,8 @@ export const FIELD_MAPPINGS: AliasMap = {
   ],
   'short_description.es': ['Program Description (Spanish)'],
   program_status: ['Program Status'],
-  program_start_raw: ['Program Start', 'Program Start (mm/dd/yyyy)'],
-  program_end_raw: ['Program End', 'Program End (mm/dd/yyyy)'],
+  start_date: ['Program Start', 'Program Start (mm/dd/yyyy)'],
+  end_date: ['Program End', 'Program End (mm/dd/yyyy)'],
   payment_methods: ['Rebate Type'],
   rebate_value: ['Rebate Value *'],
   'amount.type': ['Amount Type *'],
